Fixed angle-based shadowcasting to check all necessary quadrants. This fixes issue #67.  Also, modified radius value in SenseMap and FOV to be a double, to support euclidean-radiuses fully.

+++ GoRogue/SenseMapping/SenseMap.cs
@@ -267,10 +267,10 @@ namespace GoRogue.SenseMapping
 		private static void blitSenseSource(SenseSource source, double[,] destination, HashSet<Coord> sourceMap, IMapView<double> resMap)
 		{
 			// Calculate actual radius bounds, given constraint based on location
-			int minX = Math.Min(source.Radius, source.Position.X);
-			int minY = Math.Min(source.Radius, source.Position.Y);
-			int maxX = Math.Min(source.Radius, resMap.Width - 1 - source.Position.X);
-			int maxY = Math.Min(source.Radius, resMap.Height - 1 - source.Position.Y);
+			int minX = Math.Min((int)source.Radius, source.Position.X);
+			int minY = Math.Min((int)source.Radius, source.Position.Y);
+			int maxX = Math.Min((int)source.Radius, resMap.Width - 1 - source.Position.X);
+			int maxY = Math.Min((int)source.Radius, resMap.Height - 1 - source.Position.Y);
 
 			// Use radius bounds to extrapalate global coordinate scheme mins and maxes
 			Coord gMin = source.Position - Coord.Get(minX, minY);
@@ -278,8 +278,8 @@ namespace GoRogue.SenseMapping
 
 			// Use radius bound to extrapalate light-local coordinate scheme min and max bounds that
 			// are actually blitted
-			Coord lMin = Coord.Get(source.Radius - minX, source.Radius - minY);
-			Coord lMax = Coord.Get(source.Radius + maxX, source.Radius + maxY);
+			Coord lMin = Coord.Get((int)source.Radius - minX, (int)source.Radius - minY);
+			Coord lMax = Coord.Get((int)source.Radius + maxX, (int)source.Radius + maxY);
 
 			for (int xOffset = 0; xOffset <= lMax.X - lMin.X; xOffset++)
 			//Parallel.For(0, lMax.X - lMin.X + 1, xOffset => // By light radius 30 or so, there is enough work to get benefit here.  Manual thread splitting may also be an option.
